// client/src/generated-ts-client.ts
export interface ApplicationUserDto {
    id?: string;
    fullName?: string;
    email?: string;
    phoneNumber?: string;
    isActivePlayer?: boolean;
    role?: string;
    createdAt?: string;
    updatedAt?: string | undefined;
}
// ...
export interface BoardDto {
    id?: string;
    gameId?: string;
    playerId?: string;
    numberCount?: number;
    priceDkk?: number;
    isWinningBoard?: boolean;
    subscriptionId?: string | undefined;
    createdAt?: string;
    updatedAt?: string | undefined;
    numbers?: number[];
    game?: GameDto | undefined;
    player?: ApplicationUserDto | undefined;
    playerName?: string | undefined;
}

export interface GameDto {
    id?: string;
    weekNumber?: number;
    year?: number;
    startDate?: string;
    endDate?: string;
    guessDeadline?: string;
    isActive?: boolean;
    winningNumber1?: number | undefined;
    winningNumber2?: number | undefined;
    winningNumber3?: number | undefined;
    numbersPublishedAt?: string | undefined;
    createdAt?: string;
    updatedAt?: string | undefined;
}